Update: text file reading and querying
In this update I've managed to fix the issue with reading from file.
I've also done a little more work on the database querying and have come
to the conclusion that some aspects of the project will need to be
talked over. By this I specifically mean the handling of the user input
since edit distance doesn't seem to work all that well.